+++ control/sound/sound.go
@@ -2,10 +2,10 @@ package sound
 
 import (
 	"errors"
-	"log"
 
 	"github.com/go-ole/go-ole"
 	"github.com/moutend/go-wca/pkg/wca"
+	"github.com/tiredsosha/warden/tools/logger"
 )
 
 func GetVolume() (uint8, error) {
@@ -17,9 +17,9 @@ func GetVolume() (uint8, error) {
 	})
 
 	if err != nil || vol == nil {
-		log.Println("can't get volume level - Windows API doesn't responce")
-		log.Println(err)
-		log.Printf("volume:%s\n", vol)
+		logger.Warn.Println("can't get volume level - Windows API doesn't responce")
+		logger.Warn.Println(err)
+		logger.Info.Printf("volume:%s\n", vol)
 		err = errors.New("WINAPI")
 		return 0, err
 	}
@@ -34,8 +34,8 @@ func SetVolume(volume int) {
 	})
 
 	if err != nil {
-		log.Println("can't set volume level - Windows API doesn't responce")
-		log.Println(err)
+		logger.Warn.Println("can't set volume level - Windows API doesn't responce")
+		logger.Warn.Println(err)
 	}
 }
 
@@ -46,8 +46,8 @@ func Mute(state bool) {
 	})
 
 	if err != nil {
-		log.Println("can't set mute - Windows API doesn't responce")
-		log.Println(err)
+		logger.Warn.Println("can't set mute - Windows API doesn't responce")
+		logger.Warn.Println(err)
 	}
 }